# docs/api-reference/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Pynomaly API",
    "version": "1.0.0",
...
    "contact": {
      "name": "Pynomaly Support",
      "url": "https://pynomaly.com/support",
      "email": "support@pynomaly.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.pynomaly.com",
      "description": "Production server"
    },
    {
      "url": "https://staging-api.pynomaly.com",
      "description": "Staging server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Get current user profile",
        "description": "Get the current authenticated user's profile",
        "operationId": "get_current_user",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/UserProfile"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Error type URI",
            "example": "https://pynomaly.com/errors/validation-error"
          },
          "title": {
            "type": "string",
            "description": "Error title",
            "example": "Validation Error"
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code",
            "example": 422
          },
          "detail": {
            "type": "string",
            "description": "Error details",
            "example": "The request data is invalid"
          },
          "instance": {
            "type": "string",
            "description": "Request instance that caused the error",
            "example": "/api/v1/detection/detect"
          }
        }
      }
    },
    "responses": {
...
      "UnauthorizedError": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "type": "https://pynomaly.com/errors/unauthorized",
              "title": "Unauthorized",
              "status": 401,
              "detail": "Authentication required"
            }
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from the authentication endpoint"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for service-to-service authentication"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyAuth": []
    }
  ],
...
}